Moved the chart to a cell in the tableview
